Separate deirectory 'test' made for unit tests, which holds separate unit tests files for each modules. 'Gulpfile.js' and 'package.json' updated accordingly
